Refactoring all classes and endpoints
