+++ src/classes/Turma.ts
@@ -9,19 +9,6 @@ export class Turma extends BaseDB {
         super()
     }
 
-    public static createTurma = async (novaTurma: Turma) => {
-        await Turma
-            .connection(Turma.tableTurma)
-            .insert(novaTurma)
-    }
-
-    public static getTurmas = async () => {
-        const result = await Turma
-            .connection(Turma.tableTurma)
-            .select()
-        return result
-    }
-
     public static changeModulo = async (id:string, value:string) => {
         await Turma
             .connection(Turma.tableTurma)
