+++ src/classes/Usuario.ts
@@ -1,6 +1,4 @@
 import { BaseDB } from "./BaseDB"
-import { Docente } from "./Docente"
-import { Estudante } from "./Estudante"
 
 export abstract class Usuario extends BaseDB {
     constructor(
@@ -18,19 +16,5 @@ export abstract class Usuario extends BaseDB {
             .connection(tableName)
             .where("id","=",idUsuario)
             .update("turma_id",idNovaTurma)
-    }
-
-    public static addUsuario = async(tableName:string, novoUsuario: Docente | Estudante) => {
-        await Usuario
-            .connection(tableName)
-            .insert(novoUsuario)
-    }
-
-    public static getUsuario = async (tableName:string, nomeUsuario:string) => {
-        const result = await Usuario
-            .connection(tableName)
-            .select()
-            .where("nome","like",`%${nomeUsuario}%`)
-        return result
     }
 }
