+++ src/endpoints/addTurma.ts
@@ -1,7 +1,7 @@
 import { Response, Request } from "express";
-import { Turma } from "../../classes/Turma";
+import { Turma } from "../classes/Turma";
 
-export const createTurma = async (req:Request, res:Response) => {
+export const addTurma = async (req:Request, res:Response) => {
     let errorCode = 400
     try {
         const {nome, modulo} = req.body
@@ -19,7 +19,7 @@ export const createTurma = async (req:Request, res:Response) => {
             modulo
         )
 
-        await Turma.createTurma(novaTurma)
+        await Turma.addEntity(Turma.tableTurma, novaTurma)
         
         res.status(201).send({message: "Turma criada", turma: novaTurma})
     } catch (error: any) {
