+++ src/endpoints/getEntity.ts
@@ -1,7 +1,7 @@
 import { Response, Request } from "express";
-import { Usuario } from "../../classes/Usuario"
+import { BaseDB } from "../classes/BaseDB";
 
-export const getUsuario = async (req:Request, res:Response) => {
+export const getEntity = async (req:Request, res:Response) => {
     let errorCode = 400
     try {
         const table = req.body.table
@@ -12,17 +12,19 @@ export const getUsuario = async (req:Request, res:Response) => {
             nome = "%"
         }
         if (!table) {
-            throw new Error("Escolha tabela estudante ou docente.")
+            throw new Error("Escolha tabela estudante, docente ou turma.")
         }
         if (table.toLowerCase() === 'estudante') {
-            tableName = Usuario.tableEstudante
+            tableName = BaseDB.tableEstudante
         } else if (table.toLowerCase() === 'docente') {
-            tableName = Usuario.tableDocente
+            tableName = BaseDB.tableDocente
+        } else if (table.toLowerCase() === 'turma') {
+            tableName = BaseDB.tableTurma
         } else {
-            throw new Error("Escolha tabela estudante ou docente.")
+            throw new Error("Escolha tabela estudante, docente ou turma.")
         }
         
-        const result = await Usuario.getUsuario(tableName, nome)
+        const result = await BaseDB.getEntity(tableName, nome)
         res.status(200).send(result)
     } catch (error: any) {
         res.status(errorCode).send({ message: error.message })
