+++ src/index.ts
@@ -1,15 +1,14 @@
 import app from "./app"
-import { changeTurma } from "./endpoints/usuario/changeTurma"
-import { addUsuario } from "./endpoints/usuario/addUsuario"
-import { getUsuario } from "./endpoints/usuario/getUsuario"
-import { changeModulo } from "./endpoints/turma/changeModulo"
-import { createTurma } from "./endpoints/turma/createTurma"
-import { getTurmas } from "./endpoints/turma/getTurmas"
+import { changeTurma } from "./endpoints/changeTurma"
+import { addUsuario } from "./endpoints/addUsuario"
+import { getEntity } from "./endpoints/getEntity"
+import { changeModulo } from "./endpoints/changeModulo"
+import { addTurma } from "./endpoints/addTurma"
 
-app.put("/usuario", changeTurma)
+app.get("/entity", getEntity)
+
 app.post("/usuario", addUsuario)
-app.get("/usuario", getUsuario)
+app.put("/usuario", changeTurma)
 
-app.get("/turma", getTurmas)
-app.post("/turma", createTurma)
+app.post("/turma", addTurma)
 app.put("/turma", changeModulo)
